use json input

+++ pomodoroplayer.go
@@ -1,66 +1,88 @@
 package main
 
 import (
+	"encoding/json"
 	"fmt"
+	"os"
 	"os/exec"
 	"time"
+	"io/ioutil"
 )
 
 const playcmd = "/usr/bin/afplay" // mac only!!
 
 type Pomodoro struct {
-	worksongs []string
-	restsongs []string
-	worktime  time.Duration
-	resttime  time.Duration
+	WorkSongs []string
+	WorkTime  time.Duration
+	RestSongs []string
+	RestTime  time.Duration
 }
 
 func NewPomodoro(
 	worksongs []string,
-	restsongs []string,
 	worktime time.Duration,
+	restsongs []string,
 	resttime time.Duration) *Pomodoro {
-	return &Pomodoro{worksongs, restsongs, worktime, resttime}
+	return &Pomodoro{worksongs, worktime, restsongs, resttime}
 }
 
 func (p *Pomodoro) Start() {
-	working(p.worktime, p.worksongs)
-	working(p.resttime, p.restsongs)
+	working(p.WorkTime, p.WorkSongs)
+	working(p.RestTime, p.RestSongs)
+	fmt.Println("Pomodoro End")
 }
 
-func play(cmd *exec.Cmd, playend chan struct{}) {
-	cmd.Run()
-	playend <- struct{}{}
+func play(cmd *exec.Cmd, ch chan string) {
+	cmd.Start()
+	cmd.Wait()
+	ch <- "playend"
+}
+
+func timer(t time.Duration, ch chan string) {
+	time.Sleep(t)
+	ch <- "timeout"
 }
 
 func working(t time.Duration, worksongs []string) {
+	playend := make(chan string)
+
+	go timer(t, playend)
+
 	var cmd *exec.Cmd
-	go func() {
-		playend := make(chan struct{})
 		for _, song := range worksongs {
 			cmd = exec.Command(playcmd, song)
 
 			fmt.Printf("Start: %s\n", song)
 			go play(cmd, playend)
-			<-playend
+
+			s := <-playend
+			if s == "timeout" {
+				cmd.Process.Kill()
+				break
+			}
 		}
-	}()
-
-	time.Sleep(t)
-	cmd.Process.Kill()
-
 }
 
 func main() {
-	pomodoro := NewPomodoro(
-		[]string{
-			"./Japanese_School_Bell02-12.mp3",
-			"./Japanese_School_Bell02-12.mp3",
-		},
-		[]string{
-			"./gup01.mp4",
-		},
-		11*time.Second,
-		5*time.Second)
+	var musiclist string
+	if len(os.Args) > 1 {
+		musiclist = os.Args[1]
+	} else {
+		fmt.Println("Usage: pmodoroplayer musiclist.json")
+		os.Exit(1)
+	}
+	raw, err := ioutil.ReadFile(musiclist)
+	if err != nil {
+		fmt.Println(err.Error())
+		os.Exit(1)
+	}
+
+	var pomodoro Pomodoro
+
+	json.Unmarshal(raw, &pomodoro)
+
+	pomodoro.WorkTime = 4*time.Minute
+	pomodoro.RestTime = 2*time.Second
+
 	pomodoro.Start()
 }
